+++ utils/error/index.d.ts
@@ -1,0 +1,10 @@
+export * from "../../lib/build/error";
+/**
+ * 'export *' does not re-export a default.
+ * import SuperTokens from "supertokens-react-native";
+ * the above import statement won't be possible unless either
+ * - user add "esModuleInterop": true in their tsconfig.json file
+ * - we do the following change:
+ */
+import * as _default from "../../lib/build/error";
+export default _default;